+++ SBVTestSuite/TestSuite/Basics/AllSat.hs
@@ -31,7 +31,7 @@ tests =
     ]
 
 srt :: AllSatResult -> AllSatResult
-srt (AllSatResult (b1, b2, rs)) = AllSatResult (b1, b2, sortOn getModelDictionary rs)
+srt (AllSatResult (b1, b2, b3, rs)) = AllSatResult (b1, b2, b3, sortOn getModelDictionary rs)
 
 newtype Q = Q () deriving (Eq, Ord, Data, Read, Show, SymVal, HasKind)
 type SQ = SBV Q